move convert to dict test to seperate file
